requirements.txt
pynput>=1.7.6
pyaudio>=0.2.11
openai-whisper>=20231117
numpy>=1.24.0
